+++ lib/cocoapods-binary-cache/pod-rome/xcodebuild_command.rb
@@ -57,6 +57,7 @@ module PodPrebuild
       args_[:default] += ["DEBUG_INFORMATION_FORMAT=dwarf"] if disable_dsym?
       args_[:simulator] += ["ARCHS=x86_64", "ONLY_ACTIVE_ARCH=NO"] if simulator == "iphonesimulator"
       args_[:simulator] += args_[:default]
+      args_[:device] += ["ONLY_ACTIVE_ARCH=NO"]
       args_[:device] += args_[:default]
       args_
     end